api/requirements.txt
fastapi
uvicorn
psycopg2
python-dotenv
skl2onnx
joblib
huggingface-hub
onnxruntime
numpy